requirements.txt

streamlit
qrcode
segno
Pillow
